+++ flask_dance/contrib/google.py
@@ -16,7 +16,7 @@ def make_google_blueprint(
         client_id=None, client_secret=None, scope=None,
         offline=False, reprompt_consent=False,
         redirect_url=None, redirect_to=None, login_url=None, authorized_url=None,
-        session_class=None, backend=None):
+        session_class=None, backend=None, hosted_domain=None):
     """
     Make a blueprint for authenticating with Google using OAuth 2. This requires
     a client ID and client secret from Google. You should either pass them to
@@ -49,6 +49,41 @@ def make_google_blueprint(
         backend: A storage backend class, or an instance of a storage
                 backend class, to use for this blueprint. Defaults to
                 :class:`~flask_dance.consumer.backend.session.SessionBackend`.
+        hosted_domain (str, optional): The domain of the G Suite user. Used to indicate that the account selection UI
+            should be optimized for accounts at this domain. Note that this only provides UI optimization, and requires
+            response validation (see warning).
+
+    .. _google_hosted_domain_warning:
+    .. warning::
+       The ``hosted_domain`` argument **only provides UI optimization**. Don't rely on this argument to control
+       who can access your application. You must verify that the ``hd`` claim of the response ID token matches the
+       ``hosted_domain`` argument passed to ``make_google_blueprint``. For example:
+
+       .. code-block:: python
+
+            from flask import session, abort
+            from flask_dance.consumer import oauth_authorized
+            from flask_dance.contrib.google import make_google_blueprint, google
+            import requests
+
+            google_bp = make_google_blueprint(
+                client_id="foo",
+                client_secret="bar",
+                scope=["profile", "email"],
+                hosted_domain="example.com"
+            )
+
+            @oauth_authorized.connect_via(google_bp)
+            def logged_in(blueprint, token):
+                resp_json = google.get("/oauth2/v2/userinfo").json()
+                if resp_json["hd"] != blueprint.authorization_url_params["hd"]:
+                    requests.post(
+                        "https://accounts.google.com/o/oauth2/revoke",
+                        params={"token": token["access_token"]}
+                    )
+                    session.clear()
+                    abort(403)
+
 
     :rtype: :class:`~flask_dance.consumer.OAuth2ConsumerBlueprint`
     :returns: A :ref:`blueprint <flask:blueprints>` to attach to your Flask app.
@@ -61,6 +96,8 @@ def make_google_blueprint(
         auto_refresh_url = "https://accounts.google.com/o/oauth2/token"
     if reprompt_consent:
         authorization_url_params["approval_prompt"] = "force"
+    if hosted_domain:
+        authorization_url_params["hd"] = hosted_domain
     google_bp = OAuth2ConsumerBlueprint("google", __name__,
         client_id=client_id,
         client_secret=client_secret,